move appTypr attr to config attr store
