+++ src/client/src/store/selection/combineReducers.ts
@@ -1,10 +1,8 @@
 import { combineReducers } from "redux";
 
 import services from "../azureProfileData";
-import appType from "./app/selectWebAppReducer";
 
 const SelectionStateReducer = combineReducers({
-  appType,
   services
 });
